Fixed fetching data, fixed searching by phone number, code refactor
