+++ packages/esm-sms-app/src/sms-logs/pagination-constant.ts
@@ -1,0 +1,3 @@
+export const DEFAULT_PAGE_SIZE = 10;
+export const DEFAULT_PAGE_SIZES = [10, 20, 50, 100];
+export const DEFAULT_PAGE_NUMBER = 1;
