+++ packages/esm-sms-app/src/sms-logs/sms-logs-table.scss
