+++ packages/esm-sms-app/translations/en.json
@@ -1,7 +1,8 @@
 {
   "add": "Add",
   "addProviderConfigWorkspaceTitle": "Add Provider",
-  "config": "Configuration",
+  "clearSearchInput": "Clear search input",
+  "config": "Config",
   "configDetails": "Configuration details",
   "configuration": "Configuration",
   "configurations": "Configurations",
@@ -12,7 +13,7 @@
   "deliveryStatus": "Delivery status",
   "edit": "Edit",
   "editConfig": "Edit {{configName}}",
-  "emptyTableStateText": "There are no {{displayText}} to display",
+  "emptyTableStateText": "There are no SMS logs to display",
   "errorMakingDefault": "Failed to update default config",
   "errorMessage": "Error message",
   "fitlerLogsByNumber": "Fitler logs by phone number",
@@ -22,16 +23,20 @@
   "messageContent": "Message",
   "modificationDate": "Modification date",
   "modifiedBy": "Modified by",
+  "nextPage": "Next page",
   "noConfigurationsToDisplay": "No configurations to display",
   "noSmsLogsToDisplay": "No SMS Logs to display",
   "openMrsId": "Openmrs ID",
   "phoneNumber": "Phone number",
-  "providerId": "Provider",
+  "previousPage": "Previous page",
+  "providerId": "Provider ID",
   "providerName": "Provider name",
   "providers": "Providers",
   "providerStatus": "Provider status",
   "record": "Record",
   "remove": "Remove",
+  "search": "Search",
+  "searchByPhoneNumber": "Search by phone number",
   "setDefault": "Set as default",
   "smsConfiguration": "SMS Configuration",
   "smsDirection": "SMS direction",
