--- packages/esm-sms-app/translations/fr.json
@@ -1,4 +1,0 @@
-{
-  "smsConfiguration": "SMS Configuration",
-  "smsProviderSettings": "SMS Provider Settings"
-}
